tests: Update all tests to include the expected output
The assert_eq_fmt macro now expects ` => "expected"` at the end, where
"expected" is whatever the print should produce. This makes it easier to
see what each test is supposed to cover.

Note that this does make the tests somewhat platform dependent, as the
expected string might differ depending on the target and
toolchain. However, in practice these tests are always being run on
`x86_64-unknown-linux-gnu` currently, and we can cfg the tests to cover
other toolchains later.